requirements.txt
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pycodestyle==2.9.1
python-dotenv==0.21.0
requests==2.28.1
robodk==5.5.0
tomli==2.0.1
urllib3==1.26.12
